+++ src/stores/widgets/SpectralProfileWidget/SpectralProfileWidgetStore.ts
@@ -22,6 +22,7 @@ type Comments = string[];
 export type MultiPlotData = {
     numProfiles: number;
     data: DataPoints[];
+    secondaryXData: number[][];
     smoothedData: DataPoints[];
     fittingData: {x: number[]; y: Float32Array | Float64Array};
     colors: string[];
@@ -62,6 +63,8 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
     @observable linePlotInitXYBoundaries: {minXVal: number; maxXVal: number; minYVal: number; maxYVal: number};
     @observable settingsTabId: SpectralProfilerSettingsTabs;
 
+    @observable secondaryAxisCursorInfoVisible: boolean;
+
     // line key will be "Primary" in single line mode
     public static readonly PRIMARY_LINE_KEY = "Primary";
 
@@ -84,6 +87,12 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
 
     @action setSpectralCoordinate = (coordStr: string) => {
         if (this.effectiveFrame.setSpectralCoordinate(coordStr)) {
+            this.clearXBounds();
+        }
+    };
+
+    @action setSpectralCoordinateSecondary = (coordStr: string) => {
+        if (this.effectiveFrame.setSpectralCoordinateSecondary(coordStr)) {
             this.clearXBounds();
         }
     };
@@ -285,6 +294,10 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
         this.settingsTabId = tabId;
     };
 
+    @action setSecondaryAxisCursorInfoVisible = (val: boolean) => {
+        this.secondaryAxisCursorInfoVisible = val;
+    };
+
     constructor(coordinate: string = "z") {
         super(RegionsType.CLOSED_AND_POINT);
         makeObservable<SpectralProfileWidgetStore, "spectralLinesMHz" | "updateRanges">(this);
@@ -295,6 +308,7 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
         // Describes how the data is visualised
         this.plotType = PlotType.STEPS;
         this.meanRmsVisible = false;
+        this.secondaryAxisCursorInfoVisible = false;
         this.markerTextVisible = false;
         this.primaryLineColor = "auto-blue";
         this.lineColorMap = new Map<LineKey, string>([[SpectralProfileWidgetStore.PRIMARY_LINE_KEY, this.primaryLineColor]]);
@@ -398,6 +412,7 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
         // Determine points/smoothingPoints/colors/yBound/progress
         let numProfiles = 0;
         let data = [];
+        let secondaryXData = [];
         let smoothedData = [];
         let colors = [];
         let labels = [];
@@ -407,9 +422,10 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
         let yMean = undefined;
         let yRms = undefined;
         let progressSum: number = 0;
-        let startEndIndexes: {startIndex: number; endIndex: number}[] = [];
+        let dataIndexes: {startIndex: number; endIndex: number}[] = [];
         const wantMeanRms = profiles.length === 1;
         const profileColorMap = this.lineColorMap;
+
         profiles.forEach(profile => {
             if (profile?.data) {
                 numProfiles++;
@@ -419,8 +435,11 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
 
                 const intensityValues = this.intensityConversion ? this.intensityConversion(profile.data.values) : profile.data.values;
                 const pointsAndProperties = this.getDataPointsAndProperties(profile.channelValues, intensityValues, wantMeanRms);
+
                 data.push(pointsAndProperties?.points ?? []);
                 smoothedData.push(pointsAndProperties?.smoothedPoints ?? []);
+                secondaryXData.push(profile.channelSecondaryValues?.slice(pointsAndProperties?.startIndex, pointsAndProperties?.endIndex + 1) ?? []);
+
                 if (pointsAndProperties) {
                     if (wantMeanRms) {
                         if (this.smoothingStore.type === SmoothingType.NONE) {
@@ -444,16 +463,16 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
                         yBound.yMax = pointsAndProperties.yBound.yMax;
                     }
                     progressSum = progressSum + profile.data.progress;
-                    startEndIndexes.push({startIndex: pointsAndProperties.startIndex, endIndex: pointsAndProperties.endIndex});
+                    dataIndexes.push({startIndex: pointsAndProperties.startIndex, endIndex: pointsAndProperties.endIndex});
                 }
             }
         });
 
         let fittingData: {x: number[]; y: Float32Array | Float64Array};
-        if (profiles.length === 1 && startEndIndexes.length === 1) {
-            let x = profiles[0].channelValues.slice(startEndIndexes[0].startIndex, startEndIndexes[0].endIndex + 1);
+        if (profiles.length === 1 && dataIndexes.length === 1) {
+            let x = profiles[0].channelValues.slice(dataIndexes[0].startIndex, dataIndexes[0].endIndex + 1);
             const intensityValues = this.intensityConversion ? this.intensityConversion(profiles[0].data.values) : profiles[0].data.values;
-            let y = intensityValues.slice(startEndIndexes[0].startIndex, startEndIndexes[0].endIndex + 1);
+            let y = intensityValues.slice(dataIndexes[0].startIndex, dataIndexes[0].endIndex + 1);
             if (this.smoothingStore.type !== SmoothingType.NONE) {
                 const smoothedData = this.smoothingStore.getSmoothingValues(x, y);
                 x = smoothedData.x;
@@ -480,6 +499,7 @@ export class SpectralProfileWidgetStore extends RegionWidgetStore {
         return {
             numProfiles: numProfiles,
             data: data,
+            secondaryXData: secondaryXData,
             smoothedData: smoothedData,
             fittingData: fittingData,
             colors: colors,